Move ModalValuesType to batch types file
We deleted all the code from `BatchesCellRenderer.tsx` other than
`ModalValuesType`, which left that type in a file with a misleading
name. Move it to the types file for batches.
