--- src/scenes/InventoryRouter/view/BatchesCellRenderer.tsx
@@ -1,5 +1,0 @@
-export type ModalValuesType = {
-  type: string;
-  openChangeQuantityModal: boolean;
-  batch?: any;
-};
