+++ src/types/Batch.ts
@@ -92,3 +92,9 @@ export const batchHistoryEventEnumToLocalized = (batchHistoryType: BatchHistoryP
     }
   }
 };
+
+export type ModalValuesType = {
+  type: string;
+  openChangeQuantityModal: boolean;
+  batch?: any;
+};
